% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ks_var_less_one.R
\name{ks_var_less_one}
\alias{ks_var_less_one}
\title{ks_var_less_one}
\usage{
ks_var_less_one(tbla, formula_model, tbla_test)
}
\arguments{
\item{tbla}{table with data. It has to have the target_name. It will be used to train the new model when deleting one variable}

\item{formula_model}{the final formula of the model.}

\item{tbla_test}{table with test data.}

\item{modelo_object}{the model trained to predict the target variable.}
}
\description{
get Komolgorov-Smirnov value test for a model in train and test, deleting one variable at a time
.
}
\examples{


}
