% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ks_var_list.R
\name{ks_var_list}
\alias{ks_var_list}
\title{ks_var_list}
\usage{
ks_var_list(tbla, formula_model, tbla_test)
}
\arguments{
\item{tbla}{table with data. It has to have the target_name.}

\item{formula_model}{the final formula of the model.}

\item{tbla_test}{table with test data.}
}
\description{
get Komolgorov-Smirnov value test for a variable in train and test(receives a list or variables)
.
}
\examples{


}
