% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vars_stability.R
\name{vars_stability}
\alias{vars_stability}
\title{vars_stability}
\usage{
vars_stability(tbla, formula_model, target_name, splitter_name)
}
\arguments{
\item{tbla}{table with data. It has to have the target_name.}

\item{target_name}{the name of the target.}

\item{splitter_name}{the name of the variable that is going to be used to split the data.}

\item{model}{the model with the list of variables.}
}
\description{
get Komolgorov-Smirnov value test for a variable for a list of periods/or other variable.
.
}
\examples{


set.seed(1)
x1 = rnorm(1000)
x2 = rnorm(1000)
x4='A'
x4=ifelse(x1>0.1,'B', x4)
x4=ifelse(x1>0.4,'C', x4 )
x4=ifelse(x1>0.6,'D', x4 )
x4=ifelse(x1>0.8,'E', x4 )
per=c(rep(201810,200),rep(201812,400),rep(201901,400))
z = 1 + 3*x1 + x2
pr = 1/(1+exp(-z))
y = rbinom(1000,1,pr)
tbla = data.frame(y=y,x1=x1,x2=x2, x4=x4, per=per)
formula_model=formula(y~x2+x1+x4)
aa=vars_stability(tbla, formula_model,'y', 'per')
aa[1]
aa[2]
aa[3]
aa[4]
}
