% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ks_var_list.R
\name{ks_var_list}
\alias{ks_var_list}
\title{ks_var_list}
\usage{
ks_var_list(tbla, formula_model, tbla_test)
}
\arguments{
\item{tbla}{table with data. It has to have the target_name.}

\item{formula_model}{the final formula of the model.}

\item{tbla_test}{table with test data.}
}
\description{
get Komolgorov-Smirnov value test for a variable in train and test(receives a list or variables)
.
}
\examples{
set.seed(1)
x1 = rnorm(1000)
x2 = rnorm(1000)
z = 1 + 3*x1
pr = 1/(1+exp(-z))
y = rbinom(1000,1,pr)
tbla = data.frame(y=y,x1=x1,x2=x2, x4=x4)
formula_model=formula('y~x1+x2')
ks_var_list(tbla, formula_model, tbla)
}
